Migrate to gradle / java 17 / groovy 4 (WiP)
* gradle setup
* java 17
* groovy 4
* use java.net.http.HttpClient instead of the unsupported groovyx.net.http.HTTPBuilder
