+++ .gitignore
@@ -2,3 +2,5 @@
 *.iml
 target/
 
+.gradle/
+build/
