+++ gradle.properties
@@ -1,0 +1,1 @@
+group=com.droidablebee.gphoto
