+++ src/main/groovy/com/droidablebee/gphoto/GooglePhotoApplication.groovy
