+++ src/main/groovy/com/droidablebee/gphoto/HttpException.groovy
