+++ src/test/groovy/com/droidablebee/gphoto/GooglePhotoServiceSpec.groovy
@@ -1,11 +1,12 @@
 package com.droidablebee.gphoto
 
-import groovyx.net.http.HTTPBuilder
+
 import groovyx.net.http.HttpResponseDecorator
 import org.apache.http.HttpResponse
 import org.apache.http.StatusLine
 import spock.lang.Specification
-import spock.lang.Unroll
+
+import java.net.http.HttpClient
 
 import static com.droidablebee.gphoto.GooglePhotoService.ALBUMS
 import static com.droidablebee.gphoto.GooglePhotoService.ALBUMS_MAX_PAGE_SIZE
@@ -26,8 +27,8 @@ import static com.droidablebee.gphoto.GooglePhotoService.QUERY
 
 class GooglePhotoServiceSpec extends Specification {
 
-    HTTPBuilder http = Mock()
-    GooglePhotoService service = new GooglePhotoService(http: http)
+    HttpClient httpClient = Mock()
+    GooglePhotoService service = new GooglePhotoService(httpClient: httpClient)
 
     HttpResponse base = Mock()
     StatusLine statusLine = Mock()
@@ -37,7 +38,7 @@ class GooglePhotoServiceSpec extends Specification {
         base.statusLine >> statusLine
     }
 
-    @Unroll
+    
     def "get all albums"() {
 
         String token = "token"
@@ -50,14 +51,14 @@ class GooglePhotoServiceSpec extends Specification {
 
 //        calls * http.get(_ as Map) >>> new HttpResponseDecorator(base, data1) >> new HttpResponseDecorator(base, data2)
 
-        call1 * http.get({ Map params ->
+        call1 * httpClient.get({ Map params ->
             params[PATH] == service.getDefaultUri() + ALBUMS &&
                     params[HEADERS][AUTHORIZATION] == "${BEARER} ${token}" &&
                     params[QUERY][PAGE_SIZE] == ALBUMS_MAX_PAGE_SIZE &&
                     params[QUERY][PAGE_TOKEN] == null
         }) >> new HttpResponseDecorator(base, data1)
 
-        call2 * http.get({ Map params ->
+        call2 * httpClient.get({ Map params ->
             params[PATH] == service.getDefaultUri() + ALBUMS &&
                     params[HEADERS][AUTHORIZATION] == "${BEARER} ${token}" &&
                     params[QUERY][PAGE_SIZE] == ALBUMS_MAX_PAGE_SIZE &&
@@ -73,7 +74,7 @@ class GooglePhotoServiceSpec extends Specification {
         200        | [[id: "1", title: "title 1", productUrl: "http://mock/1"]] | [[id: "2", title: "title 2", productUrl: "http://mock/2"]] | 1     | 1     | [(ALBUMS): albums1, (NEXT_PAGE_TOKEN): "nextPageToken"] | [(ALBUMS): albums2]
     }
 
-    @Unroll
+    
     def "get all albums - error"() {
 
         String token = "token"
@@ -88,7 +89,7 @@ class GooglePhotoServiceSpec extends Specification {
         exception.code == error.code
         exception.status == error.status
 
-        1 * http.get(_ as Map) >> response
+        1 * httpClient.get(_ as Map) >> response
         1 * statusLine.statusCode >> statusCode
 
         where:
@@ -96,7 +97,7 @@ class GooglePhotoServiceSpec extends Specification {
         401        | [code: statusCode, message: "Request had invalid authentication credentials", status: "UNAUTHENTICATED"] | [(ERROR): error]
     }
 
-    @Unroll
+    
     def "get items for album"() {
 
         String token = "token"
@@ -110,7 +111,7 @@ class GooglePhotoServiceSpec extends Specification {
 
 //        calls * http.post(_ as Map) >>> new HttpResponseDecorator(base, data1) >> new HttpResponseDecorator(base, data2)
 
-        call1 * http.post({ Map params ->
+        call1 * httpClient.post({ Map params ->
             params[PATH] == service.getDefaultUri() + MEDIA_ITEMS_SEARCH &&
                     params[HEADERS][AUTHORIZATION] == "${BEARER} ${token}" &&
                     params[BODY][ALBUM_ID] == albumId &&
@@ -118,7 +119,7 @@ class GooglePhotoServiceSpec extends Specification {
                     params[BODY][PAGE_TOKEN] == null
         }) >> new HttpResponseDecorator(base, data1)
 
-        call2 * http.post({ Map params ->
+        call2 * httpClient.post({ Map params ->
             params[PATH] == service.getDefaultUri() + MEDIA_ITEMS_SEARCH &&
                     params[HEADERS][AUTHORIZATION] == "${BEARER} ${token}" &&
                     params[BODY][ALBUM_ID] == albumId &&
@@ -135,7 +136,7 @@ class GooglePhotoServiceSpec extends Specification {
         200        | [[id: "1", description: "description 1", productUrl: "http://mock/1"]] | [[id: "2", description: "description 2", productUrl: "http://mock/2"]] | 1     | 1     | [(MEDIA_ITEMS): items1, (NEXT_PAGE_TOKEN): "nextPageToken"] | [(MEDIA_ITEMS): items2]
     }
 
-    @Unroll
+    
     def "get items for album - error"() {
 
         String token = "token"
@@ -151,7 +152,7 @@ class GooglePhotoServiceSpec extends Specification {
         exception.code == error.code
         exception.status == error.status
 
-        1 * http.post(_ as Map) >> response
+        1 * httpClient.post(_ as Map) >> response
         1 * statusLine.statusCode >> statusCode
 
         where:
@@ -159,7 +160,7 @@ class GooglePhotoServiceSpec extends Specification {
         401        | [code: statusCode, message: "Request had invalid authentication credentials", status: "UNAUTHENTICATED"] | [(ERROR): error]
     }
 
-    @Unroll
+    
     def "get all items"() {
 
         String token = "token"
@@ -172,14 +173,14 @@ class GooglePhotoServiceSpec extends Specification {
 
 //        calls * http.get(_ as Map) >>> new HttpResponseDecorator(base, data1) >> new HttpResponseDecorator(base, data2)
 
-        call1 * http.get({ Map params ->
+        call1 * httpClient.get({ Map params ->
             params[PATH] == service.getDefaultUri() + MEDIA_ITEMS &&
                     params[HEADERS][AUTHORIZATION] == "${BEARER} ${token}" &&
                     params[QUERY][PAGE_SIZE] == ITEMS_MAX_PAGE_SIZE &&
                     params[QUERY][PAGE_TOKEN] == null
         }) >> new HttpResponseDecorator(base, data1)
 
-        call2 * http.get({ Map params ->
+        call2 * httpClient.get({ Map params ->
             params[PATH] == service.getDefaultUri() + MEDIA_ITEMS &&
                     params[HEADERS][AUTHORIZATION] == "${BEARER} ${token}" &&
                     params[QUERY][PAGE_SIZE] == ITEMS_MAX_PAGE_SIZE &&
@@ -195,7 +196,7 @@ class GooglePhotoServiceSpec extends Specification {
         200        | [[id: "1", description: "description 1", productUrl: "http://mock/1"]] | [[id: "2", description: "description 2", productUrl: "http://mock/2"]] | 1     | 1     | [(MEDIA_ITEMS): items1, (NEXT_PAGE_TOKEN): "nextPageToken"] | [(MEDIA_ITEMS): items2]
     }
 
-    @Unroll
+    
     def "get all items - error"() {
 
         String token = "token"
@@ -210,7 +211,7 @@ class GooglePhotoServiceSpec extends Specification {
         exception.code == error.code
         exception.status == error.status
 
-        1 * http.get(_ as Map) >> response
+        1 * httpClient.get(_ as Map) >> response
         1 * statusLine.statusCode >> statusCode
 
         where:
